requirements.txt
pandas
omadb
numpy
jupyter
biopython
bytecode
selenium
matplotlib
requests
torch
openpyxl
plotly
tqdm
joblib
sklearn
comet-ml
webdriver-manager
jellyfish
py-cpuinfo
kaleido
GitPython
git+https://github.com/luo3300612/Visualizer
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
